Replace substanceregistry with substancesSystem
