+++ Assets/Scripts/SS3D/Systems/Substances/SubstancesSystem.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 25179695ba5be5e49b80bce55c54d185
+guid: 16647fb19ee14c44e979af85557daf59
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
